requirements.txt
ansimarkup==1.4.0
better-exceptions-fork==0.2.1.post6
colorama==0.4.1
cx-Oracle==7.1.3
grpcio==1.20.0
grpcio-tools==1.20.0
loguru==0.2.5
protobuf==3.7.1
Pygments==2.3.1
six==1.12.0
